+++ models/db/ConsultationAgendaItem.php
@@ -176,10 +176,9 @@ class ConsultationAgendaItem extends ActiveRecord
             $childItem->deleteWithAllDependencies();
         }
 
-        foreach ($this->getMyMotions() as $motion) {
-            $motion->agendaItemId = null;
-            $motion->save();
-        }
+        // We use SQL queries here to make sure that no foreign key from any soft-deleted motion/amendment still points to this agenda item, making it undeletable.
+        \Yii::$app->db->createCommand('UPDATE `motion` SET `agendaItemId` = NULL WHERE `agendaItemId` = ' . intval($this->id))->execute();
+        \Yii::$app->db->createCommand('UPDATE `amendment` SET `agendaItemId` = NULL WHERE `agendaItemId` = ' . intval($this->id))->execute();
 
         $this->delete();
     }